# Conflicts:
#	src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/template/v2/UserPipelineTemplateV2Resource.kt
#	src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/constant/ProcessMessageCode.kt
#	src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/pipeline/PipelineResourceOnlyVersion.kt
#	src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/pipeline/PipelineResourceVersion.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/template/v2/UserPipelineTemplateV2ResourceImpl.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/PipelineVersionFacadeService.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/pipeline/version/PipelineVersionGenerator.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/pipeline/version/convert/PipelineDraftSaveReqConverter.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/PipelineTemplateGenerator.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/PipelineTemplatePersistenceService.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/PipelineTemplateResourceService.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/version/convert/PipelineTemplateDraftRollbackReqConverter.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/version/convert/PipelineTemplateDraftSaveReqConverter.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/version/hander/PipelineTemplateDraftSaveHandler.kt
#	src/backend/ci/core/process/biz-process/src/test/kotlin/com/tencent/devops/process/service/template/v2/PipelineTemplateGeneratorTest.kt

+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/provider/rbac/service/RbacPermissionResourceService.kt
@@ -308,24 +308,32 @@ class RbacPermissionResourceService(
                 resourceCode = resourceCode
             )
             if (resourceInfo != null) {
-                val projectInfo = authResourceService.get(
-                    projectCode = projectCode,
-                    resourceType = AuthResourceType.PROJECT.value,
-                    resourceCode = projectCode
-                )
-                val deleteTime = DateTimeUtil.toDateTime(LocalDateTime.now(), "yyMMddHHmmSS")
-                val deleteResourceName = "${resourceInfo.resourceName}[$deleteTime]"
-                // 权限中心删除是异步删除,当删除后,立马创建重名资源,权限中心会报资源冲突,所以需要先修改资源名称后删除
-                permissionSubsetManagerService.modifySubsetManager(
-                    subsetManagerId = resourceInfo.relationId,
-                    projectCode = projectCode,
-                    projectName = projectInfo.resourceName,
-                    resourceType = resourceType,
-                    resourceCode = resourceCode,
-                    resourceName = deleteResourceName,
-                    iamResourceCode = resourceInfo.iamResourceCode
-                )
-                permissionSubsetManagerService.deleteSubsetManager(resourceInfo.relationId)
+                // 资源已禁用且relationId为空,说明二级管理员已在禁用时被删除,无需再调用IAM删除接口
+                if (!resourceInfo.enable && resourceInfo.relationId.isBlank()) {
+                    logger.info(
+                        "resource is disabled and relationId is empty, " +
+                            "skip delete subset manager|$projectCode|$resourceType|$resourceCode"
+                    )
+                } else {
+                    val projectInfo = authResourceService.get(
+                        projectCode = projectCode,
+                        resourceType = AuthResourceType.PROJECT.value,
+                        resourceCode = projectCode
+                    )
+                    val deleteTime = DateTimeUtil.toDateTime(LocalDateTime.now(), "yyMMddHHmmSS")
+                    val deleteResourceName = "${resourceInfo.resourceName}[$deleteTime]"
+                    // 权限中心删除是异步删除,当删除后,立马创建重名资源,权限中心会报资源冲突,所以需要先修改资源名称后删除
+                    permissionSubsetManagerService.modifySubsetManager(
+                        subsetManagerId = resourceInfo.relationId,
+                        projectCode = projectCode,
+                        projectName = projectInfo.resourceName,
+                        resourceType = resourceType,
+                        resourceCode = resourceCode,
+                        resourceName = deleteResourceName,
+                        iamResourceCode = resourceInfo.iamResourceCode
+                    )
+                    permissionSubsetManagerService.deleteSubsetManager(resourceInfo.relationId)
+                }
             }
         }
         authResourceService.delete(
@@ -408,8 +416,36 @@ class RbacPermissionResourceService(
             logger.info("resource has enable permission manager|$userId|$projectId|$resourceType|$resourceCode")
             return true
         }
+        // 如果发现resourceInfo.relationId不为空，说明二级管理员已存在，不需要再创建
+        val subsetManagerId: Int
+        val createMode: AuthGroupCreateMode
+        if (resourceInfo.relationId.isNotBlank()) {
+            subsetManagerId = resourceInfo.relationId.toInt()
+            createMode = AuthGroupCreateMode.ENABLE
+        } else {
+            // 重新创建二级管理员
+            subsetManagerId = permissionSubsetManagerService.createSubsetManager(
+                gradeManagerId = projectInfo.relationId,
+                userId = userId,
+                projectCode = projectId,
+                projectName = projectInfo.resourceName,
+                resourceType = resourceType,
+                resourceCode = resourceCode,
+                resourceName = resourceInfo.resourceName,
+                iamResourceCode = resourceInfo.iamResourceCode
+            )
+            createMode = AuthGroupCreateMode.CREATE
+            // 更新资源的二级管理员关联ID
+            authResourceService.updateRelationId(
+                projectCode = projectId,
+                resourceType = resourceType,
+                resourceCode = resourceCode,
+                relationId = subsetManagerId.toString()
+            )
+        }
+        // 创建默认用户组
         permissionSubsetManagerService.createSubsetManagerDefaultGroup(
-            subsetManagerId = resourceInfo.relationId.toInt(),
+            subsetManagerId = subsetManagerId,
             userId = userId,
             projectCode = projectId,
             projectName = projectInfo.resourceName,
@@ -417,7 +453,7 @@ class RbacPermissionResourceService(
             resourceCode = resourceCode,
             resourceName = resourceInfo.resourceName,
             iamResourceCode = resourceInfo.iamResourceCode,
-            createMode = AuthGroupCreateMode.ENABLE
+            createMode = createMode
         )
         return authResourceService.enable(
             userId = userId,
